+++ human-verification/presentation/src/test/kotlin/me/proton/core/humanverification/presentation/viewmodel/hv3/HV3ViewModelTest.kt
@@ -41,7 +41,7 @@ import me.proton.core.network.domain.session.SessionId
 import me.proton.core.test.kotlin.CoroutinesTest
 import me.proton.core.usersettings.domain.entity.RecoverySetting
 import me.proton.core.usersettings.domain.entity.UserSettings
-import me.proton.core.usersettings.domain.usecase.GetSettings
+import me.proton.core.usersettings.domain.usecase.GetUserSettings
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -54,7 +54,7 @@ class HV3ViewModelTest : CoroutinesTest {
     private val humanVerificationWorkflowHandler = mockk<HumanVerificationWorkflowHandler>(relaxed = true)
     private val humanVerificationListener = mockk<HumanVerificationListener>(relaxed = true)
     private val accountRepository = mockk<AccountRepository>(relaxed = true)
-    private val getSettings = mockk<GetSettings>(relaxed = true)
+    private val getSettings = mockk<GetUserSettings>(relaxed = true)
     private val networkPrefs = mockk<NetworkPrefs>(relaxed = true)
 
     lateinit var viewModel: HV3ViewModel
@@ -114,7 +114,7 @@ class HV3ViewModelTest : CoroutinesTest {
             every { locale } returns "en_US"
             every { phone } returns RecoverySetting("123456789", 0, false, false)
         }
-        coEvery { getSettings.invoke(any()) } returns settingsMock
+        coEvery { getSettings.invoke(any(), any()) } returns settingsMock
         val params = viewModel.getHumanVerificationExtraParams()
         assertNotNull(params)
         assertEquals("US", params?.defaultCountry)
@@ -137,7 +137,7 @@ class HV3ViewModelTest : CoroutinesTest {
             every { locale } returns "en_US"
             every { phone } returns RecoverySetting("123456789", 0, false, false)
         }
-        coEvery { getSettings.invoke(any()) } returns settingsMock
+        coEvery { getSettings.invoke(any(), any()) } returns settingsMock
         val params = viewModel.getHumanVerificationExtraParams()
         assertEquals(true, params?.useVPNTheme)
     }